+++ config/medium/ASCIICube/code/CGCode.php
@@ -1,0 +1,9 @@
+<?php
+fscanf(STDIN, "%d", $w);
+fscanf(STDIN, "%d", $h);
+fscanf(STDIN, "%d", $d);
+
+// Write an answer using echo(). DON'T FORGET THE TRAILING \n
+// To debug: error_log(var_export($var, true)); (equivalent to var_dump)
+
+echo("ASCII cube\n");